almost

+++ main/Dashboard/views.py
@@ -2,28 +2,30 @@ from django.shortcuts import render,HttpResponse,redirect
 from django.contrib.auth.models import User
 from django.contrib import messages
 from Dashboard.models import Profile,Registered_Books,Issued_Books,Returned_Books
+from django.contrib.auth.decorators import login_required
+# from django.contrib.admin.views.decorators import staff_member_required
 from Authentication.send_mail import send_mail
 from Authentication.models import ChangePassword
 from uuid import uuid4
+from django.utils import timezone
 
 # Create your views here.
 def index(request):
     title = '''Home -'''
     if request.user.is_authenticated :
-        print("in authenticated")
         title = '''Dashboard-'''
         staff_status = "User"       #DEfault
         book_obj = Registered_Books.objects.all().count()
         if (request.user.is_superuser) :
             staff_status = "Admin"
-            print("in superUser")
             profile_count = Profile.objects.all().count()
             return render(request,"index.html",{"title":title,"staff_status":staff_status,"registeredUsers":profile_count,"bookListedNumbers":book_obj})
 
         return render(request,"index.html",{"title":title,"staff_status":staff_status,"bookListedNumbers":book_obj})
-
-    print("in home")
-    return render(request,'index.html',{"title":title})
+    
+    else:
+        messages.add_message(request, messages.WARNING, "Login First !!!")
+        return render(request,'index.html',{"title":title})
 
 def regStudents(request):
     title = '''Registered Students'''
@@ -32,6 +34,8 @@ def regStudents(request):
 
     return render(request,'registered-students.html',{"title":title,"students":students})
 
+
+@login_required(login_url="http://127.0.0.1:8000/")
 def editProfile(request,myid):
     title = '''Edit-Profile'''
     profile = Profile.objects.get(library_id = myid)
@@ -60,7 +64,11 @@ def editProfile(request,myid):
             return redirect('http://127.0.0.1:8000/edit-profile/{}'.format(myid))
     return render(request,"profile.html",{"title":title,"profile":profile})
 
+
 def delprofile(request,myid):
+    if not request.user.is_superuser:
+        messages.add_message(request, messages.WARNING, "Admin Login First !!!")
+        return redirect('http://127.0.0.1:8000/authenticate/admin-login/')
     profile = Profile.objects.filter(id = myid)
     # profile = Profile.objects.filter(id = myid).delete()
     profile.delete()
@@ -96,7 +104,9 @@ def forgotPassword(request):
 
 
 def addBook(request):
-
+    if not request.user.is_superuser:
+        messages.add_message(request, messages.WARNING, "Admin Login First !!!")
+        return redirect('http://127.0.0.1:8000/authenticate/admin-login/')
     title = '''adding book'''
     try:
         if request.method == 'POST':
@@ -129,25 +139,37 @@ def addBook(request):
 
 
 def deleteBook(request,isbn):
+    if not request.user.is_superuser:
+        messages.add_message(request, messages.WARNING, "Admin Login First !!!")
+        return redirect('http://127.0.0.1:8000/authenticate/admin-login/')
     book_obj = Registered_Books.objects.get(ISBN = isbn).delete()
 
     return redirect('http://127.0.0.1:8000/manage-book/')
 
 def deleteProfile(request,myid):
+    if not request.user.is_superuser:
+        messages.add_message(request, messages.WARNING, "Admin Login First !!!")
+        return redirect('http://127.0.0.1:8000/authenticate/admin-login/')
     profile = Profile.objects.get(library_id = myid)
     
     User.objects.get(id = profile.user.id).delete()
 
     return redirect('http://127.0.0.1:8000/registered-students/')
 
+@login_required(login_url='http://127.0.0.1:8000/')
 def listedBooks(request):
     if request.user.is_authenticated:
         title = '''Registered Books'''
         book_obj = Registered_Books.objects.all()
+        print(book_obj[0].status)
         
         return render(request,'registered-books.html',{"title":title,"books":book_obj})
 
+
 def manageBook(request):
+    if not request.user.is_superuser:
+        messages.add_message(request, messages.WARNING, "Admin Login First !!!")
+        return redirect('http://127.0.0.1:8000/authenticate/admin-login/')
     title = '''Manage Book'''
     if request.user.is_superuser:
         title = '''Registered Books'''
@@ -157,6 +179,10 @@ def manageBook(request):
 
 
 def bookDetails(request,isbn):
+    if not request.user.is_superuser:
+        messages.add_message(request, messages.WARNING, "Admin Login First !!!")
+        return redirect('http://127.0.0.1:8000/authenticate/admin-login/')
+    
     title = '''Update Books Details'''
     print(isbn)
     book_obj1 = Registered_Books.objects.get(ISBN = isbn)
@@ -193,6 +219,8 @@ def bookDetails(request,isbn):
     
     return render(request,'book-detail.html', {"title":title,"book":book_obj1})
 
+
+@login_required(login_url='http://127.0.0.1:8000/')
 def searchBy(request):
     title = '''Registered Books'''
     if request.method == 'POST':
@@ -207,18 +235,160 @@ def searchBy(request):
         print(union_queryset)
 
         return render(request,'registered-books.html',{"title":title,"books":union_queryset})
-    
+
+
+@login_required(login_url='http://127.0.0.1:8000/')  
 def applyFilter(request):
     title = '''Registered Books'''
     if request.method == 'POST':
         sort_by = request.POST.get('sort-by')
         
         book_obj1 = Registered_Books.objects.all().order_by(sort_by)
-        # print(book_obj1)
-        # book_obj2 = Registered_Books.objects.filter(ISBN__contains = search)
-        # book_obj3 = Registered_Books.objects.filter(authorName__contains = search)
-        # book_obj4 = Registered_Books.objects.filter(department__contains = search)
-        # book_obj5 = Registered_Books.objects.filter(category__contains = search)
-        # union_queryset = book_obj1 | book_obj2 | book_obj3 | book_obj4 | book_obj5
-
         return render(request,'registered-books.html',{"title":title,"books":book_obj1})
+    
+
+def issueBook(request):
+    print(" in issue book funtion")
+    if not request.user.is_superuser:
+        messages.add_message(request, messages.WARNING, "Admin Login First !!!")
+        return redirect('http://127.0.0.1:8000/authenticate/admin-login/')
+    title = '''Issue Book'''
+
+    try:
+        if request.method == 'POST':
+            print("posted")
+            ISBN = request.POST.get('ISBN')
+            library_id = request.POST.get('library_id')
+            print(ISBN,library_id)
+            found = True
+            book_obj = Registered_Books.objects.filter(ISBN = ISBN)
+            profile_obj = Profile.objects.filter(library_id = library_id)
+            if not book_obj.exists():
+                found = False
+                messages.add_message(request, messages.WARNING, "Book Not Found !!!")
+
+            if not profile_obj.exists():
+                found = False
+                messages.add_message(request, messages.WARNING, "Both Book and Student Not Found !!!")
+
+            return render(request,'issue-book.html',{"title":title,"book":book_obj,"student":profile_obj,"found":found})
+
+    except Exception as e:
+        return redirect('/')
+
+    
+
+def issuingBook(request):
+    if not request.user.is_superuser:
+        messages.add_message(request, messages.WARNING, "Admin Login First !!!")
+        return redirect('http://127.0.0.1:8000/authenticate/admin-login/')
+    
+    try:
+        if request.method == 'POST':
+            ISBN = request.POST.get('ISBN')
+            library_id = request.POST.get('library_id')
+            
+
+            book_obj = Registered_Books.objects.get(ISBN = ISBN)
+            profile_obj = Profile.objects.get(library_id = library_id)   
+
+            profile_obj.issuedBooks = profile_obj.issuedBooks + 1
+            book_obj.last_issued_by = profile_obj.name
+            print(book_obj.status)
+            book_obj.status = "Not Available"
+            print(book_obj.status)
+
+            issued_book_obj = Issued_Books.objects.create(issued_by = profile_obj.name, email = profile_obj.user.email, mobile = profile_obj.mobile, issue_date = timezone.now().strftime('%Y-%m-%d'),library_id = profile_obj.library_id,bookName = book_obj.bookName, authorName = book_obj.authorName, department = book_obj.department, ISBN = book_obj.ISBN, category = book_obj.category )
+
+            book_obj.save()
+            print(book_obj.status)
+            profile_obj.save()
+            issued_book_obj.save()
+
+            return redirect('http://127.0.0.1:8000/issued-books/')
+
+    except Exception as e:
+        print(e)
+        messages.add_message(request, messages.SUCCESS, "Some Error !!")
+        return redirect('http://127.0.0.1:8000/')
+    
+    
+def returnBook(request):
+    print(" in return book funtion")
+    if not request.user.is_superuser:
+        messages.add_message(request, messages.WARNING, "Admin Login First !!!")
+        return redirect('http://127.0.0.1:8000/authenticate/admin-login/')
+    title = '''Issue Book'''
+
+    try:
+        if request.method == 'POST':
+            print("posted")
+            ISBN = request.POST.get('ISBN')
+            library_id = request.POST.get('library_id')
+            print(ISBN,library_id)
+            found = True
+            book_obj = Registered_Books.objects.filter(ISBN = ISBN)
+            profile_obj = Profile.objects.filter(library_id = library_id)
+            if not book_obj.exists():
+                found = False
+                messages.add_message(request, messages.WARNING, "Book Not Found !!!")
+
+            if not profile_obj.exists():
+                found = False
+                messages.add_message(request, messages.WARNING, "Both Book and Student Not Found !!!")
+
+            return render(request,'return-book.html',{"title":title,"book":book_obj,"student":profile_obj,"found":found})
+
+    except Exception as e:
+        return redirect('/')
+    
+
+def returningBook(request):
+    if not request.user.is_superuser:
+        messages.add_message(request, messages.WARNING, "Admin Login First !!!")
+        return redirect('http://127.0.0.1:8000/authenticate/admin-login/')
+    
+    try:
+        if request.method == 'POST':
+            ISBN = request.POST.get('ISBN')
+            library_id = request.POST.get('library_id')
+
+            book_obj = Registered_Books.objects.get(ISBN = ISBN)
+            profile_obj = Profile.objects.get(library_id = library_id)   
+
+            profile_obj.returnedBooks = profile_obj.returnedBooks + 1
+            print(book_obj.status)
+            Registered_Books.objects.update(last_issued_by = profile_obj.name, status = 'Available')
+            print(book_obj.status)
+
+            returned_books_obj = Returned_Books.objects.create(returned_by = profile_obj.name, email = profile_obj.user.email, mobile = profile_obj.mobile, return_date = timezone.now().strftime('%Y-%m-%d'),library_id = profile_obj.library_id,bookName = book_obj.bookName, authorName = book_obj.authorName, department = book_obj.department, ISBN = book_obj.ISBN, category = book_obj.category )
+
+            book_obj.save()
+            print(book_obj.status)
+            profile_obj.save()
+            returned_books_obj.save()
+
+            return redirect('http://127.0.0.1:8000/returned-books/')
+
+    except Exception as e:
+        print(e)
+        messages.add_message(request, messages.SUCCESS, "Some Error !!")
+        return redirect('http://127.0.0.1:8000/')
+    
+def issuedBooks(request):
+    if not request.user.is_superuser:
+        messages.add_message(request, messages.WARNING, "Admin Login First !!!")
+        return redirect('http://127.0.0.1:8000/authenticate/admin-login/')
+    title = '''BBAU-LIBRARY | Issued Books'''
+    issued_book_obj = Returned_Books.objects.all().count()
+
+    return render(request,'issued-books.html',{"title":title,"issuedBooksNumbers":issued_book_obj})
+    
+def returnedBooks(request):
+    if not request.user.is_superuser:
+        messages.add_message(request, messages.WARNING, "Admin Login First !!!")
+        return redirect('http://127.0.0.1:8000/authenticate/admin-login/')
+    title = '''BBAU-LIBRARY | Returned Books'''
+    returned_books_obj = Returned_Books.objects.all().count()
+
+    return render(request,'returned-books.html',{"title":title,"returnedBooksNumbers":returned_books_obj})
